+++ spyne/application.py
@@ -101,8 +101,17 @@ class Application(object):
         self.in_protocol = in_protocol
         self.out_protocol = out_protocol
 
-        self.in_protocol.set_app(self)
-        self.out_protocol.set_app(self)
+        if self.in_protocol is None:
+            from spyne.protocol import ProtocolBase
+            self.in_protocol = ProtocolBase(self)
+        else:
+            self.in_protocol.set_app(self)
+
+        if self.out_protocol is None:
+            from spyne.protocol import ProtocolBase
+            self.out_protocol = ProtocolBase(self)
+        else:
+            self.out_protocol.set_app(self)
 
         register_application(self)