+++ spyne/protocol/xml/model.py
@@ -30,6 +30,7 @@ from spyne.const.xml_ns import soap_env as _ns_soap_env
 from spyne.error import Fault
 from spyne.error import ValidationError
 from spyne.model import PushBase
+from spyne.model.complex import XmlData
 from spyne.model.complex import XmlAttribute
 from spyne.util import coroutine
 from spyne.util import Break
@@ -185,6 +186,10 @@ def get_members_etree(prot, cls, inst, parent):
                     v.marshall(prot, k, subvalue, parent)
                 continue
 
+            elif issubclass(v, XmlData):
+                v.marshall(prot, k, subvalue, parent)
+                continue
+
             mo = v.Attributes.max_occurs
             if subvalue is not None and mo > 1:
                 if isinstance(subvalue, PushBase):
@@ -268,6 +273,11 @@ def complex_from_element(prot, cls, element):
 
     # this is for validating cls.Attributes.{min,max}_occurs
     frequencies = defaultdict(int)
+
+    xtba_key, xtba_type = cls.Attributes._xml_tag_body_as
+    if xtba_key is not None:
+        value = prot.from_string(xtba_type.type, element.text)
+        setattr(inst, xtba_key, value)
 
     # parse input to set incoming data to related attributes.
     for c in element: